+++ src/prime_rl/trainer/models/afmoe/modeling_afmoe.py
@@ -351,13 +351,7 @@ class AfmoeDecoderLayer(GradientCheckpointingLayer):
 
         residual = hidden_states
         hidden_states = self.pre_mlp_layernorm(hidden_states)
-        if isinstance(self.mlp, MoE):
-            hidden_states = self.mlp(
-                hidden_states,
-                routed_experts=routed_experts,
-            )
-        else:
-            hidden_states = self.mlp(hidden_states)
+        hidden_states = self.mlp(hidden_states, routed_experts=routed_experts)
         hidden_states = self.post_mlp_layernorm(hidden_states)
         hidden_states = residual + hidden_states
         return hidden_states
