+++ src/prime_rl/trainer/models/glm4_moe/modeling_glm4_moe.py
@@ -121,13 +121,7 @@ class Glm4MoeDecoderLayer(GradientCheckpointingLayer):
         # Fully Connected
         residual = hidden_states
         hidden_states = self.post_attention_layernorm(hidden_states)
-        if isinstance(self.mlp, MoE):
-            hidden_states = self.mlp(
-                hidden_states,
-                routed_experts=routed_experts,
-            )
-        else:
-            hidden_states = self.mlp(hidden_states)
+        hidden_states = self.mlp(hidden_states, routed_experts=routed_experts)
         hidden_states = residual + hidden_states
         return hidden_states
